requirements.txt
crispy-forms-materialize==0.2
Django==2.1.1
django-crispy-forms==1.7.2
mysqlclient==1.3.13
Pillow==5.3.0
pytz==2018.5
